requirements.txt
dateparser==1.0.0
discord.py==1.7.3
python-dotenv==0.19.0